Move int-set to read_fonts::collections::intset
This is gated on 'std'; if we want to be finer-grained we can look into
that as followup.
