+++ Cargo.toml
@@ -11,7 +11,6 @@ members = [
     "skrifa",
     "fauntlet",
     "klippa",
-    "int-set",
     "fuzz",
 ]
