--- int-set/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "int-set"
-version = "0.1.0"
-edition = "2021"
-license = "MIT/Apache-2.0"
-description = "A fast sparse and invertible bit set for u32's. Port of harfbuzz's hb_set_t."
-
-[dependencies]
-font-types = { version = "0.5.5", path = "../font-types"}
-
-[dev-dependencies]
-criterion = "0.5.1"
-rand = "0.8.5"
-
-[[bench]]
-name = "int_set_benchmark"
-harness = false
-
-[[bench]]
-name = "sparse_bit_set_benchmark"
-harness = false
