+++ read-fonts/benches/sparse_bit_set_benchmark.rs
@@ -2,8 +2,7 @@ mod bench_helper;
 
 use bench_helper::random_set;
 use criterion::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
-use int_set::sparse_bit_set::to_sparse_bit_set_with_bf;
-use int_set::IntSet;
+use read_fonts::collections::int_set::{sparse_bit_set::to_sparse_bit_set_with_bf, IntSet};
 
 struct SparseSetTest {
     set_size: u32,
