+++ read-fonts/src/collections.rs
@@ -1,0 +1,4 @@
+//! Data structures useful for font work.
+
+pub mod int_set;
+pub use int_set::IntSet;
